Prepend clan tags to requested player name
It may be better to store this off and display seperately from the nickname,
but since the UID is used for identification, this should not cause
regressions, and it conveniently puts the clantag everywhere (lobby, in-game,
replay).

I followed current practice for parsing commandline arguments, but this code is in need of refactoring.

Closes #95