Abort CPU benchmark if the user leaves the lobby
(Text copied from pull request)

An edgecase, but it's been annoying me: The CPU benchmark runs to completion
even if the user returns to the main menu (which, admittedly, can only happen
if you're in a LAN game).

While I'd dearly like to just exterminate this ridiculous benchmark entirely,
this should go some way to making it less irritating.

More generally, we should perhaps consider replacing it with something that
gives a more accurate picture of how well a machine will cope in-game. Testing
the performance of Lua arithmetic operators isn't a sensible way to do this:
this is not the nature of the majority of the game's load.  Realistically, the
best way to produce a rating would be just to read system information (whatever
windows' equivalent of lspci, /proc/cpuinfo and /proc/meminfo are) and make a
decision based on that. As-is, the benchmark often spits out slightly stupid
numbers that aren't really representative of how well a machine will perform
in-game, and causes many seconds of UI unusability while it's thinking about
it.

Closes #243